Review fixes - Service private methods are in separate utils file now - itemsPage param renamed to just page
