+++ src/services/new/structure-item/utils.ts
@@ -1,0 +1,260 @@
+import {ServiceArgs} from '../types';
+import {getReplica} from '../utils';
+import {raw} from 'objection';
+import Utils from '../../../utils';
+import {WorkbookModel, WorkbookModelColumn} from '../../../db/models/new/workbook';
+import {CollectionModel, CollectionModelColumn} from '../../../db/models/new/collection';
+import {AppContext, AppError} from '@gravity-ui/nodekit';
+import {registry} from '../../../registry';
+import {isWorkbookInstance} from '../../../registry/common/entities/structure-item/types';
+import {Feature, isEnabledFeature} from '../../../components/features';
+import {WorkbookPermission} from '../../../entities/workbook';
+import {CollectionPermission} from '../../../entities/collection';
+import {US_ERRORS} from '../../../const';
+import {CollectionInstance} from '../../../registry/common/entities/collection/types';
+import {WorkbookInstance} from '../../../registry/common/entities/workbook/types';
+
+interface GetWorkbooksQueryArgs {
+    collectionId: Nullable<string>;
+    filterString?: string;
+    onlyMy?: boolean;
+}
+export const getWorkbooksQuery = ({ctx, trx}: ServiceArgs, args: GetWorkbooksQueryArgs) => {
+    const {filterString, onlyMy, collectionId} = args;
+    const {
+        tenantId,
+        projectId,
+        user: {userId},
+    } = ctx.get('info');
+
+    const targetTrx = getReplica(trx);
+    return WorkbookModel.query(targetTrx)
+        .select({
+            type: raw("'workbook'"),
+            workbookId: WorkbookModelColumn.WorkbookId,
+            collectionId: WorkbookModelColumn.CollectionId,
+            title: WorkbookModelColumn.Title,
+            sortTitle: WorkbookModelColumn.SortTitle,
+            description: WorkbookModelColumn.Description,
+            parentId: raw('null'),
+            projectId: WorkbookModelColumn.ProjectId,
+            tenantId: WorkbookModelColumn.TenantId,
+            createdBy: WorkbookModelColumn.CreatedBy,
+            createdAt: WorkbookModelColumn.CreatedAt,
+            updatedBy: WorkbookModelColumn.UpdatedBy,
+            updatedAt: WorkbookModelColumn.UpdatedAt,
+            meta: WorkbookModelColumn.Meta,
+        })
+        .where({
+            [WorkbookModelColumn.TenantId]: tenantId,
+            [WorkbookModelColumn.ProjectId]: projectId,
+            [WorkbookModelColumn.CollectionId]: collectionId,
+            [WorkbookModelColumn.DeletedAt]: null,
+        })
+        .where((qb) => {
+            if (filterString) {
+                const preparedFilterString = Utils.escapeStringForLike(filterString.toLowerCase());
+                qb.where(WorkbookModelColumn.TitleLower, 'LIKE', `%${preparedFilterString}%`);
+            }
+            if (onlyMy) {
+                qb.where({
+                    [WorkbookModelColumn.CreatedBy]: userId,
+                });
+            }
+        });
+};
+
+interface GetCollectionsQueryArgs {
+    collectionId: Nullable<string>;
+    filterString?: string;
+    onlyMy?: boolean;
+}
+export const getCollectionsQuery = ({ctx, trx}: ServiceArgs, args: GetCollectionsQueryArgs) => {
+    const {filterString, onlyMy, collectionId} = args;
+    const {
+        tenantId,
+        projectId,
+        user: {userId},
+    } = ctx.get('info');
+
+    const targetTrx = getReplica(trx);
+
+    return CollectionModel.query(targetTrx)
+        .select({
+            type: raw("'collection'"),
+            workbookId: raw('null'),
+            collectionId: CollectionModelColumn.CollectionId,
+            title: CollectionModelColumn.Title,
+            sortTitle: CollectionModelColumn.SortTitle,
+            description: CollectionModelColumn.Description,
+            parentId: CollectionModelColumn.ParentId,
+            projectId: CollectionModelColumn.ProjectId,
+            tenantId: CollectionModelColumn.TenantId,
+            createdBy: CollectionModelColumn.CreatedBy,
+            createdAt: CollectionModelColumn.CreatedAt,
+            updatedBy: CollectionModelColumn.UpdatedBy,
+            updatedAt: CollectionModelColumn.UpdatedAt,
+            meta: CollectionModelColumn.Meta,
+        })
+        .where({
+            [CollectionModelColumn.TenantId]: tenantId,
+            [CollectionModelColumn.ProjectId]: projectId,
+            [CollectionModelColumn.DeletedAt]: null,
+            [CollectionModelColumn.ParentId]: collectionId,
+        })
+        .where((qb) => {
+            if (filterString) {
+                const preparedFilterString = Utils.escapeStringForLike(filterString.toLowerCase());
+                qb.where(CollectionModelColumn.TitleLower, 'LIKE', `%${preparedFilterString}%`);
+            }
+            if (onlyMy) {
+                qb.where({
+                    [CollectionModelColumn.CreatedBy]: userId,
+                });
+            }
+        });
+};
+
+interface ProcessPermissionsArgs {
+    ctx: AppContext;
+    models: (CollectionModel | WorkbookModel)[];
+    parentIds: string[];
+    skipCheckPermissions: boolean;
+    includePermissionsInfo: boolean;
+}
+export const processPermissions = async ({
+    ctx,
+    models,
+    parentIds,
+    skipCheckPermissions,
+    includePermissionsInfo,
+}: ProcessPermissionsArgs) => {
+    let result;
+
+    const {accessServiceEnabled} = ctx.config;
+    const {Workbook, Collection} = registry.common.classes.get();
+
+    if (accessServiceEnabled && !skipCheckPermissions) {
+        const checkedItems = await Promise.all(
+            models.map(async (model: WorkbookModel | CollectionModel) => {
+                const item = isWorkbookModel(model)
+                    ? new Workbook({ctx, model})
+                    : new Collection({ctx, model});
+
+                try {
+                    if (isWorkbookInstance(item)) {
+                        await item.checkPermission({
+                            parentIds,
+                            permission: isEnabledFeature(ctx, Feature.UseLimitedView)
+                                ? WorkbookPermission.LimitedView
+                                : WorkbookPermission.View,
+                        });
+                    } else {
+                        await item.checkPermission({
+                            parentIds,
+                            permission: isEnabledFeature(ctx, Feature.UseLimitedView)
+                                ? CollectionPermission.LimitedView
+                                : CollectionPermission.View,
+                        });
+                    }
+
+                    return item;
+                } catch (error) {
+                    const err = error as AppError;
+
+                    if (err.code === US_ERRORS.ACCESS_SERVICE_PERMISSION_DENIED) {
+                        return null;
+                    }
+
+                    throw error;
+                }
+            }),
+        );
+
+        result = checkedItems.filter((item) => item !== null) as InstanceType<
+            typeof Collection | typeof Workbook
+        >[];
+
+        if (includePermissionsInfo) {
+            result = await bulkFetchStructureItemsAllPermissions(
+                ctx,
+                result.map((item) => ({
+                    model: item.model,
+                    parentIds,
+                })),
+            );
+        }
+    } else {
+        result = models.map((model: WorkbookModel | CollectionModel) => {
+            const item = isWorkbookModel(model)
+                ? new Workbook({ctx, model})
+                : new Collection({ctx, model});
+
+            if (includePermissionsInfo) {
+                item.enableAllPermissions();
+            }
+
+            return item;
+        });
+    }
+    return result;
+};
+
+const isWorkbookModel = (model: CollectionModel | WorkbookModel): model is WorkbookModel => {
+    return 'workbookId' in model && Boolean(model.workbookId);
+};
+const isWorkbookItem = (item: {
+    model: CollectionModel | WorkbookModel;
+    parentIds: string[];
+}): item is {model: WorkbookModel; parentIds: string[]} => {
+    return isWorkbookModel(item.model);
+};
+const isCollectionItem = (item: {
+    model: CollectionModel | WorkbookModel;
+    parentIds: string[];
+}): item is {model: CollectionModel; parentIds: string[]} => {
+    return !isWorkbookItem(item);
+};
+
+const bulkFetchStructureItemsAllPermissions = async (
+    ctx: AppContext,
+    items: {model: CollectionModel | WorkbookModel; parentIds: string[]}[],
+) => {
+    if (items.length === 0) {
+        return [];
+    }
+
+    const {bulkFetchCollectionsAllPermissions, bulkFetchWorkbooksAllPermissions} =
+        registry.common.functions.get();
+
+    const collectionItems: {model: CollectionModel; parentIds: string[]}[] = [];
+    const workbookItems: {model: WorkbookModel; parentIds: string[]}[] = [];
+    items.forEach((item) => {
+        if (isWorkbookItem(item)) {
+            workbookItems.push(item);
+        } else if (isCollectionItem(item)) {
+            collectionItems.push(item);
+        }
+    });
+
+    const [collectionsWithPermissions, workbooksWithPermissions] = await Promise.all([
+        bulkFetchCollectionsAllPermissions(ctx, collectionItems),
+        bulkFetchWorkbooksAllPermissions(ctx, workbookItems),
+    ]);
+
+    const collectionsById: {[k: string]: CollectionInstance} = {};
+    const workbooksById: {[k: string]: WorkbookInstance} = {};
+
+    collectionsWithPermissions.forEach((item) => {
+        collectionsById[item.model.collectionId] = item;
+    });
+    workbooksWithPermissions.forEach((item) => {
+        workbooksById[item.model.workbookId] = item;
+    });
+
+    return items.map((item) => {
+        return isWorkbookModel(item.model)
+            ? workbooksById[item.model.workbookId]
+            : collectionsById[item.model.collectionId];
+    });
+};
